Sealed WindowsNotificationPart
