+++ flutter_local_notifications_windows/lib/src/details.dart
@@ -3,12 +3,10 @@ export 'details/notification_action.dart';
 export 'details/notification_audio.dart';
 export 'details/notification_details.dart';
 export 'details/notification_header.dart';
-export 'details/notification_image.dart';
 export 'details/notification_input.dart';
-export 'details/notification_part.dart';
+export 'details/notification_parts.dart';
 export 'details/notification_progress.dart';
 export 'details/notification_row.dart';
-export 'details/notification_text.dart';
 
 /// The result of updating a notification.
 enum NotificationUpdateResult {
