+++ flutter_local_notifications_windows/lib/src/details/xml/image.dart
@@ -1,6 +1,6 @@
 import 'package:xml/xml.dart';
 
-import '../notification_image.dart';
+import '../notification_parts.dart';
 
 /// Converts a [WindowsImage] to XML
 extension ImageToXml on WindowsImage {
